Refactor image cache and add unit tests
